requirements.txt
facenet_pytorch
Flask
numpy
opencv_contrib_python>=4.5.5.62
opencv_python>=4.5.5.62
opencv_python_headless>=4.5.5.62
torch>=1.11.0
torchvision>=0.12.0
waitress>=2.1.2
librosa
moviepy
google-genai
python-dotenv
